refactor(core): split cache.rs and export.rs into per-concern submodules (#180, #181)
Continues the context-burn reduction refactor series (see #184).

cache.rs (1,867 lines) → cache/{mod,config,stats,operations,preloader}.rs
export.rs (1,817 lines) → export/{mod,csv,opml,taskpaper,ical,markdown}.rs

No behavior changes; all public APIs and feature gates unchanged.
